refactor(platform): move gitdiff into platform/gitdiff
Move-only: gitdiff has zero internal imports and two unrelated consumers - the
agent's utils and the TUI's diff viewer. That is a platform leaf, not a service.

This was the second edge making internal/agent import upward. internal/services
is now empty and removed; the package no longer exists.
