+++ internal/platform/gitdiff/gitdiff.go
