+++ internal/platform/gitdiff/gitdiff_test.go
