+++ internal/presentation/tui/app/chat.go
@@ -25,6 +25,7 @@ import (
 	convdomain "github.com/inference-gateway/cli/internal/conversation/domain"
 	constants "github.com/inference-gateway/cli/internal/platform/constants"
 	formatting "github.com/inference-gateway/cli/internal/platform/formatting"
+	gitdiff "github.com/inference-gateway/cli/internal/platform/gitdiff"
 	logger "github.com/inference-gateway/cli/internal/platform/logger"
 	storage "github.com/inference-gateway/cli/internal/platform/storage"
 	shortcuts "github.com/inference-gateway/cli/internal/presentation/shortcuts"
@@ -35,7 +36,6 @@ import (
 	handlers "github.com/inference-gateway/cli/internal/presentation/tui/handlers"
 	keybinding "github.com/inference-gateway/cli/internal/presentation/tui/keybinding"
 	styles "github.com/inference-gateway/cli/internal/presentation/tui/styles"
-	gitdiff "github.com/inference-gateway/cli/internal/services/gitdiff"
 )
 
 // actChatFocusAttachments is the chat-namespace action that moves key focus to
