+++ internal/presentation/tui/components/diff_viewer.go
@@ -13,8 +13,8 @@ import (
 	tea "charm.land/bubbletea/v2"
 
 	config "github.com/inference-gateway/cli/config"
+	gitdiff "github.com/inference-gateway/cli/internal/platform/gitdiff"
 	styles "github.com/inference-gateway/cli/internal/presentation/tui/styles"
-	gitdiff "github.com/inference-gateway/cli/internal/services/gitdiff"
 )
 
 const (
